backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pytesseract
Pillow
deep-translator
python-dotenv
